move navigatorsettings to streaminfo
